feat: move shared parts to new shared class, add tests
