+++ lib/src/web/zxing_wasm/zxing_wasm_result.dart
@@ -1,10 +1,11 @@
 import 'dart:js_interop';
 import 'dart:ui';
 
-import 'package:mobile_scanner/src/enums/barcode_format.dart';
 import 'package:mobile_scanner/src/enums/barcode_type.dart';
 import 'package:mobile_scanner/src/objects/barcode.dart';
 import 'package:mobile_scanner/src/objects/barcode_bytes.dart';
+import 'package:mobile_scanner/src/web/web_barcode_utils.dart';
+import 'package:mobile_scanner/src/web/zxing_wasm/zxing_wasm_formats.dart';
 import 'package:web/web.dart' as web;
 
 /// JS interop for the `window.ZXingWASM` global set by the zxing-wasm IIFE
@@ -140,66 +141,8 @@ extension ZXingWasmReadResultToBarcode on ZXingWasmReadResult {
       rawDecodedBytes:
           rawBytes != null ? DecodedBarcodeBytes(bytes: rawBytes) : null,
       rawValue: text,
-      size: _computeSize(corners),
+      size: computeBoundingBoxSize(corners),
       type: BarcodeType.text,
     );
   }
-
-  Size _computeSize(List<Offset> corners) {
-    if (corners.length != 4) return Size.zero;
-    final xs = corners.map((c) => c.dx);
-    final ys = corners.map((c) => c.dy);
-    return Size(
-      xs.reduce((a, b) => a > b ? a : b) - xs.reduce((a, b) => a < b ? a : b),
-      ys.reduce((a, b) => a > b ? a : b) - ys.reduce((a, b) => a < b ? a : b),
-    );
-  }
-}
-
-/// Maps a zxing-wasm format string to [BarcodeFormat].
-extension ZXingWasmFormatStringToBarcodeFormat on String {
-  /// Converts a zxing-wasm format string (e.g. `'QRCode'`) to a
-  /// [BarcodeFormat] enum value.
-  BarcodeFormat get toBarcodeFormat => switch (this) {
-    'Aztec' => BarcodeFormat.aztec,
-    'Codabar' => BarcodeFormat.codabar,
-    'Code39' => BarcodeFormat.code39,
-    'Code93' => BarcodeFormat.code93,
-    'Code128' => BarcodeFormat.code128,
-    'DataMatrix' => BarcodeFormat.dataMatrix,
-    'EAN-8' => BarcodeFormat.ean8,
-    'EAN-13' => BarcodeFormat.ean13,
-    'ITF' => BarcodeFormat.itf,
-    'PDF417' => BarcodeFormat.pdf417,
-    'QRCode' => BarcodeFormat.qrCode,
-    'UPC-A' => BarcodeFormat.upcA,
-    'UPC-E' => BarcodeFormat.upcE,
-    _ => BarcodeFormat.unknown,
-  };
-}
-
-/// Maps a [BarcodeFormat] to a zxing-wasm format string, or `null` if the
-/// format is not supported by zxing-wasm.
-extension BarcodeFormatToZXingWasmString on BarcodeFormat {
-  /// Converts a [BarcodeFormat] to a zxing-wasm format string
-  /// (e.g. `'QRCode'`), or `null` if the format is not supported.
-  String? get toZXingWasmString => switch (this) {
-    BarcodeFormat.aztec => 'Aztec',
-    BarcodeFormat.codabar => 'Codabar',
-    BarcodeFormat.code39 => 'Code39',
-    BarcodeFormat.code93 => 'Code93',
-    BarcodeFormat.code128 => 'Code128',
-    BarcodeFormat.dataMatrix => 'DataMatrix',
-    BarcodeFormat.ean8 => 'EAN-8',
-    BarcodeFormat.ean13 => 'EAN-13',
-    BarcodeFormat.itf ||
-    BarcodeFormat.itf14 ||
-    BarcodeFormat.itf2of5 ||
-    BarcodeFormat.itf2of5WithChecksum => 'ITF',
-    BarcodeFormat.pdf417 => 'PDF417',
-    BarcodeFormat.qrCode => 'QRCode',
-    BarcodeFormat.upcA => 'UPC-A',
-    BarcodeFormat.upcE => 'UPC-E',
-    BarcodeFormat.all || BarcodeFormat.unknown || _ => null,
-  };
 }
